refactor(web): add server-auth shell for getSession
Introduce server-auth.ts as the single server-side session boundary while
keeping utils.ts as a thin re-export layer for minimal import churn.
