+++ apps/web/src/lib/auth/server-auth.ts
@@ -1,0 +1,72 @@
+import "server-only";
+
+import { headers } from "next/headers";
+import { redirect } from "next/navigation";
+import { cache } from "react";
+
+import { auth, type Session } from "@/lib/auth/auth";
+
+export type { Session };
+
+interface GetSessionOptions {
+  refresh?: boolean;
+}
+
+async function fetchSession(
+  requestHeaders: Headers,
+  options?: GetSessionOptions,
+): Promise<Session | null> {
+  if (options?.refresh) {
+    return auth.api.getSession({
+      query: {
+        disableCookieCache: true,
+      },
+      headers: requestHeaders,
+    });
+  }
+
+  return auth.api.getSession({
+    headers: requestHeaders,
+  });
+}
+
+const getCachedSession = cache(async (): Promise<Session | null> => {
+  return fetchSession(await headers());
+});
+
+/**
+ * Gets the current user's session information. This function only works with
+ * session-based authentication, not API keys.
+ *
+ * @returns Promise resolving to the user's session if valid, null otherwise
+ */
+export async function getSession(
+  options?: GetSessionOptions,
+): Promise<Session | null> {
+  if (options?.refresh) {
+    return fetchSession(await headers(), options);
+  }
+
+  return getCachedSession();
+}
+
+/**
+ * Gets the current user's session or redirects to the login page if no valid session is found.
+ * This is useful for protecting routes that require session-based authentication.
+ *
+ * @returns Promise resolving to the user's session if authenticated
+ * @throws {NextError} Redirects to login page with return URL when not authenticated
+ */
+export async function getSessionOrRedirect(): Promise<Session> {
+  const session = await getSession();
+  if (session) {
+    return session;
+  }
+  // Get the current URL from headers for server-side redirect
+  const headersList = await headers();
+  const pathname = headersList.get("x-pathname") ?? "";
+  const searchParams = headersList.get("x-search-params") ?? "";
+  const currentUrl = pathname + searchParams;
+  const returnUrl = encodeURIComponent(currentUrl);
+  redirect(`/signin?returnUrl=${returnUrl}`);
+}
